refactor: gallery 파일명 수정
