+++ src/components/domain/chat/ChatRoom.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useRef } from 'react';
-import Gallery from '../../../assets/chat/gallery_tmp.svg'
+import Gallery from '../../../assets/chat/gallery.svg'
 import QuotationCard from './QuotationCard';
 import PaymentCard from './PaymentCard';
 import PaymentModal, { type PaymentRequestData } from './PaymentModal';
